fix(onboarding): clear seed-flow flags on back and fix detekt
Move similar-shows fetch into its own file to stay under TooManyFunctions, clear suggestion-flow flags when leaving search/OPML suggestions, and cover isError.
