+++ feature/onboarding/src/main/java/cx/aswin/boxlore/feature/onboarding/OnboardingSuggestionsPresentation.kt
@@ -114,5 +114,8 @@ internal object OnboardingSuggestionsPresentation {
             aiLoadingStage = AiLoadingStage.IDLE,
             onboardingError = null,
             suggestionSeedCount = 0,
+            reachedSuggestionsViaSearchFlow = false,
+            reachedSuggestionsViaOpmlFlow = false,
+            reachedSuggestionsViaAiFlow = false,
         )
 }
